refactor a bit for persitence. clean up imports
