+++ package.json
@@ -30,8 +30,7 @@
     "pg": "^8.11.3",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.8.1",
-    "typeorm": "^0.2.32",
-    "uuid": "^9.0.1"
+    "typeorm": "^0.2.32"
   },
   "devDependencies": {
     "@nestjs/cli": "^10.0.0",
