+++ src/tasks/dto/get-tasks-filter.dto.ts
@@ -1,5 +1,5 @@
 import { IsEnum, IsOptional, IsString } from 'class-validator';
-import { TaskStatus } from '../tasks.model';
+import { TaskStatus } from '../tasks-status-enum';
 
 export class GetTasksFilterDto {
   @IsOptional()
