+++ src/tasks/tasks-status-enum.ts
@@ -1,9 +1,3 @@
-export interface Task {
-  id: string;
-  title: string;
-  description: string;
-  status: TaskStatus;
-}
 export enum TaskStatus {
   TODO = 'TODO',
   OPEN = 'OPEN',
